warning fixes

+++ filter-mimedefang.pl
@@ -377,7 +377,10 @@ sub data_save {
             $rh->{$hkey}{pos} //= 0;
             $rh->{$hkey}{val} = $3;
             my $hln = percent_decode($hkey) . ': ' . percent_decode( $rh->{$hkey}{val} );
-            splice @endlines, percent_decode($rh->{$hkey}{pos} - 1), 0, $hln;
+            if($rh->{$hkey}{pos} > 0) {
+              $rh->{$hkey}{pos}--;
+            }
+            splice @endlines, percent_decode($rh->{$hkey}{pos}), 0, $hln;
         }
         if ( $lfr =~ /^N([a-z\-]+)\s+([0-9]+)\s+(.*)/i ) {
             my $hkey = $1;
@@ -385,7 +388,10 @@ sub data_save {
             $rh->{$hkey}{pos} //= 0;
             $rh->{$hkey}{val} = $3;
             my $hln = percent_decode($hkey) . ': ' . percent_decode( $rh->{$hkey}{val} );
-            splice @endlines, percent_decode($rh->{$hkey}{pos} - 1), 0, $hln;
+            if($rh->{$hkey}{pos} > 0) {
+              $rh->{$hkey}{pos}--;
+            }
+            splice @endlines, percent_decode($rh->{$hkey}{pos}), 0, $hln;
         }
         if ( $lfr =~ /^J([a-z\-]+)\s+([0-9]+)/i ) {
             my $hkey = $1;
@@ -512,6 +518,6 @@ sub cleanup {
     my $message = $state->{message};
 
     if(not $debug) {
-      rmtree( $message->{md_spool_dir} );
+      rmtree( $message->{md_spool_dir} ) if defined $message->{md_spool_dir};
     }
 }
